Fix Crash on Pet Load Aura. Pet Aura is also uint32, like any other aura (same on TC2, mangos and mangos one)

+++ src/game/Pet.cpp
@@ -1900,7 +1900,7 @@ void Pet::CastPetAuras(bool current)
 
 void Pet::CastPetAura(PetAura const* aura)
 {
-    uint16 auraId = aura->GetAura(GetEntry());
+    uint32 auraId = aura->GetAura(GetEntry());
     if (!auraId)
         return;
